Fixing minipolicybox
MiniPolicyBox now tries to get the path for each policy, checks if there's
a policy at that path, and if there is, passes state in the link; if not,
it passes undefined so the policy page reloads. This eliminates the issues
with minipoliycbox passing paths through state that don't point to valid
policies.